FIX: Server Exits due to Possible Name Collision after Map Change
RCA: unique player name for clients was generated in handleUserConnected() which made sure the name was unique considering players already present in m_mapPlayers. However, adding the new player to m_mapPlayers was being done in handleUserSetupFromServer().
In a sunny weather scenario any user connecting to the server generated a handleUserConnected() and handleUserSetupFromServer() sequence in this order.
However, with multiple players reconnecting to the server in burst fashion, 2 MsgUserConnected messages might had been injected after each other into server's msg queue, that could lead to thinking both names are unique even though they were the same.

FIX: fix was to move the code responsible for unique names into handleUserNameChange() which definitely happens for any user after the MsgUserConnected and MsgUserSetupFromServer messages have been already processed for that user.
This also requires allowing a player to be inserted into m_mapPlayers with empty name.
From now on we insert a player into gameMode in MsgUserNameChange because that is the exact point when names become unique.