FIX: Mapcycle First Map when CVAR_SV_MAP is Forced
When a map was set in CVAR_SV_MAP, then if server changed the map, it loaded the 2nd map in mapcycle instead of the first. Now it loads the first. This is achieved by this trick: if something is set in CVAR_SV_MAP, then serverDecideFirstMapAndUpdateNextMapToBeLoaded() fast-forwards the mapcycle to the last map, so in case of map change mapcycleNext() will go to the first map instead of the 2nd.