FIX: Player Jump-Crouch into Tunnel
In map_warhouse, player could not go into the tunnel on the left side of the map, even tho was crouching during jump, it did not fit in.
This was happening since the commit that changed the starting Y position of map blocks in Maps::load() from 4.f to 0.f. Dont know the real reason, I guess it was always just a matter of luck, and now due to the nature of floating point calculations it did not fit in. So as a remedy I decreased player crouching height a bit.