FTR: Weapon Auto-Switch
Different cases when still has unmag ammo, and when has no unmag ammo left.
Not tested thoroughly.
Also, should avoid doing any auto stuff when player is dead.

Also, when only the current weapon is left per configured setting (did not find better to switch to), then still auto reload should be invoked. However, it is not invoked since it is either auto-reload or auto-switch. Somehow auto-reload should be triggered when no better weapon found. Dont know how to trigger it because next best weapon finding happens in InputHandling. Probably with a trick it should set the flag for WeaponHandling so in next iteration InputHandling will trigger auto-reload as it was requested by WeaponHandling. This is very tricky.