FIX: Skipped Auto-Behavior if Holding Trigger while Weapon is Becoming Empty
If client kept holding the trigger of the weapon while it became empty, the auto-behaviors were not actually triggered since server rejected them (as expected because neither reload nor switch can be executed when trigger is being held).
Solution is to re-schedule such auto-behaviors in InputHandling if trigger is held (fire button is hold), so whenever user releases it, the re-scheduled auto-behavior will be executed.