FIX: Possible Segfault with Weapon Auto-Switch
If the auto-switch is induced by pickup while the game window is inactive, the auto-switch is postponed/rescheduled until the window becomes active again.
However, if the auto-switch is not happening due to the window staying inactive, but a map change also happens in the meantime, the ptr holding the weapon becomes invalid. This is not detected because still not using smart ptrs. After the map change, when window becomes active again, the schedule wpn auto-switch is triggered, with the invalid ptr, leading to unspecified behavior, including crash.
As we are still not using smart ptrs to detect this (m_pWpnAutoSwitchWhenPickedUp should be weak_ptr), the fix now is to manually clear any scheduled switch in case of disconnect() which is also triggered in case of map change.